Abstracted all code into seperate files for you <3
